Removed code to assign recipe.
