+++ scripts/populate_batch_id.py
@@ -4,7 +4,7 @@ from datetime import date
 from EPPs.common import StepEPP
 
 
-class PopulateBatchIDRecipe(StepEPP):
+class PopulateBatchID(StepEPP):
 
     def create_cst_batch_id(self):
         today = str(date.today())
@@ -17,20 +17,10 @@ class PopulateBatchIDRecipe(StepEPP):
         raise ValueError('Cannot allocate more than 999 cst batch IDs with date %s ' % today)
 
     def _run(self):
-        #assign CST Batch ID
+        # assign CST Batch ID
         self.process.analytes()[0][0].container.name = self.create_cst_batch_id()
         self.process.analytes()[0][0].container.put()
 
-        #change the recipe to pooling if any input artifacts are pools
-        recipe='HiSeqXRecipe.xml'
-        for art in self.artifacts:
-            if art.container.name[-3:] == 'PDP':
-                recipe='HiSeqXRecipe-Pooled.xml'
-
-        #set the recipe in the relevant step UDF
-
-        self.process.udf['Sequencing Run Configuration'] = recipe
-        self.process.put()
 
 if __name__ == "__main__":
-    PopulateBatchIDRecipe().run()
+    PopulateBatchID().run()
